# server/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Backend",
        "version": "0.1.0"
    },
    "paths": {
...
        "/prompt/{prompt_id}": {
            "get": {
                "tags": [
                    "db"
                ],
                "summary": "Get Prompt",
                "operationId": "get_prompt",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "prompt_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Prompt Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Prompt"
                                }
                            }
                        }
                    },
                    "401": {
                        "detail": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "detail": "document not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "db"
                ],
                "summary": "Delete Prompt",
                "operationId": "delete_prompt",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "prompt_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Prompt Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "detail": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "detail": "document not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "db"
                ],
                "summary": "Patch Prompt",
                "operationId": "patch_prompt",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "prompt_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Prompt Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/PromptMessage"
                                },
                                "title": "Messages"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Prompt"
                                }
                            }
                        }
                    },
                    "401": {
                        "detail": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "detail": "document not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/project": {
            "post": {
                "tags": [
                    "db"
                ],
                "summary": "Post Project",
                "operationId": "post_project",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectRouteInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "detail": "Not authenticated"
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "detail": "document already exists"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "detail": "document not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/project/{project_id}": {
            "get": {
                "tags": [
                    "db"
                ],
                "summary": "Get Project",
                "operationId": "get_project",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Project Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "401": {
                        "detail": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "detail": "document not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "db"
                ],
                "summary": "Delete Project",
                "operationId": "delete_project",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Project Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "detail": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "detail": "document not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpExceptionModel"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "EvaluateResult": {
                "properties": {
                    "response": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ProviderResponse"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "vars": {
                        "anyOf": [
                            {
                                "additionalProperties": {
                                    "type": "string"
                                },
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Vars"
                    },
                    "score": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Score"
                    },
                    "latencyMs": {
                        "type": "integer",
                        "title": "Latencyms"
                    },
                    "gradingResult": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/GradingResult"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "namedScores": {
                        "additionalProperties": {
                            "type": "number"
                        },
                        "type": "object",
                        "title": "Namedscores"
                    },
                    "cost": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cost"
                    },
                    "metadata": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Metadata"
                    }
                },
                "type": "object",
                "required": [
                    "score",
                    "latencyMs",
                    "namedScores"
                ],
                "title": "EvaluateResult"
            },
            "EvaluateStats": {
                "properties": {
                    "successes": {
                        "type": "integer",
                        "title": "Successes"
                    },
                    "failures": {
                        "type": "integer",
                        "title": "Failures"
                    },
                    "tokenUsage": {
                        "$ref": "#/components/schemas/TokenUsage"
                    }
                },
                "type": "object",
                "required": [
                    "successes",
                    "failures",
                    "tokenUsage"
                ],
                "title": "EvaluateStats"
            },
            "EvaluateSummary": {
                "properties": {
                    "timestamp": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Timestamp"
                    },
                    "results": {
                        "items": {
                            "$ref": "#/components/schemas/EvaluateResult"
                        },
                        "type": "array",
                        "title": "Results"
                    },
                    "stats": {
                        "$ref": "#/components/schemas/EvaluateStats"
                    }
                },
                "type": "object",
                "required": [
                    "timestamp",
                    "results",
                    "stats"
                ],
                "title": "EvaluateSummary"
            },
            "GradingResult": {
                "properties": {
                    "pass": {
                        "type": "boolean",
                        "title": "Pass"
                    },
                    "score": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Score"
                    },
                    "componentResults": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/GradingResult"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Componentresults"
                    },
                    "assertion": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Assertion"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "comment": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Comment"
                    }
                },
                "type": "object",
                "required": [
                    "pass",
                    "score"
                ],
                "title": "GradingResult"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "HttpExceptionModel": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "status": {
                        "type": "integer",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": [
                    "message",
                    "status"
                ],
                "title": "HttpExceptionModel"
            },
...
            "Project": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 40,
                        "minLength": 1,
                        "title": "Name",
                        "examples": [
                            "My Project"
                        ]
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 1,
                        "title": "Description",
                        "examples": [
                            "This is a project"
                        ]
                    },
                    "prompt_ids": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Prompt Ids",
                        "examples": [
                            []
                        ]
                    },
                    "_id": {
                        "type": "string",
                        "title": " Id"
                    },
                    "username": {
                        "type": "string",
                        "format": "email",
                        "title": "Username"
                    },
                    "creation_time": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Creation Time"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "description",
                    "prompt_ids",
                    "username",
                    "creation_time"
                ],
                "title": "Project"
            },
            "ProjectRouteInput": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 40,
                        "minLength": 1,
                        "title": "Name",
                        "examples": [
                            "My Project"
                        ]
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 1,
                        "title": "Description",
                        "examples": [
                            "This is a project"
                        ]
                    },
                    "prompt_ids": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Prompt Ids",
                        "examples": [
                            []
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "description",
                    "prompt_ids"
                ],
                "title": "ProjectRouteInput"
            },
            "Prompt": {
                "properties": {
                    "messages": {
                        "items": {
                            "$ref": "#/components/schemas/PromptMessage"
                        },
                        "type": "array",
                        "title": "Messages",
                        "examples": [
                            [
                                {
                                    "content": "Summarize the following text: {{text}} in {{number_of_sentences}} sentences.",
                                    "role": "user"
                                }
                            ]
                        ]
                    },
                    "ai_function_id": {
                        "type": "string",
                        "title": "Ai Function Id"
                    },
                    "_id": {
                        "type": "string",
                        "title": " Id"
                    },
                    "username": {
                        "type": "string",
                        "format": "email",
                        "title": "Username"
                    },
                    "creation_time": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Creation Time"
                    },
                    "last_eval": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/EvaluateSummary"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "ai_function_name": {
                        "type": "string",
                        "title": "Ai Function Name"
                    },
                    "revision_required": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Revision Required",
                        "default": false
                    }
                },
                "type": "object",
                "required": [
                    "messages",
                    "ai_function_id",
                    "username",
                    "creation_time",
                    "ai_function_name"
                ],
                "title": "Prompt"
            },
            "PromptMessage": {
                "properties": {
                    "role": {
                        "$ref": "#/components/schemas/RoleEnum"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content"
                    }
                },
                "type": "object",
                "required": [
                    "role",
                    "content"
                ],
                "title": "PromptMessage"
            },
...
            "RoleEnum": {
                "type": "string",
                "enum": [
                    "system",
                    "user",
                    "assistant"
                ],
                "title": "RoleEnum"
            },
            "SuccessResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "Success"
                    }
                },
                "type": "object",
                "title": "SuccessResponse"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/auth/login"
                    }
                }
            }
        }
    }
}